% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed2gtf.R
\name{bed2gtf}
\alias{bed2gtf}
\title{Convert gtf file to bed12 file.}
\usage{
bed2gtf(bed_file, gtf_file = NULL)
}
\arguments{
\item{bed_file}{A bed file name to save.}

\item{gtf_file}{A gtf file.}
}
\value{
A gtf file will be saved to the directory.
}
\description{
Convert gtf file to bed12 file.
}
