% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countSliding.R
\name{countSliding}
\alias{countSliding}
\title{Count the TSS or TES number in sliding window}
\usage{
countSliding(x, wsize = 11, type = c("TTS", "TES"))
}
\arguments{
\item{x}{A data frame with a column of "TSS" or "TES" and a column of "count" shows the number
of TSSs or ETSs.
TES     count
80077         2
80078         2
80079         3
80081         4
80082        18}

\item{type}{One of "TSS" and "TES".}

\item{w}{A odd number of window size for sliding.}
}
\description{
Count the TSS or TES number in sliding window
}
