% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNArevComplement.R
\name{DNArevComplement}
\alias{DNArevComplement}
\title{Reverse complement of DNA}
\usage{
DNArevComplement(s)
}
\arguments{
\item{exons}{A string or a vector of strings}
}
\value{
A reversed complement string or vector of strings
}
\description{
Reverse complement of DNA
}
