% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon2intron.R
\name{exon2intron}
\alias{exon2intron}
\title{Get introns from exon coordinates}
\usage{
exon2intron(exons, tx2gene = NULL, add_label = T, sorted = T)
}
\arguments{
\item{exons}{A GRanges object of exons}

\item{tx2gene}{A data frame of transcript-gene association}

\item{add_label}{Logical, whether to add label of seqnames:start-end;strand}
}
\value{
A GRanges object of introns
}
\description{
Get introns from exon coordinates
}
