% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_gtf.R
\name{export_gtf}
\alias{export_gtf}
\title{Export Granges object to gtf file}
\usage{
export_gtf(
  gr,
  file2save,
  source.col = "source",
  feature.col = "feature",
  score = ".",
  phase = ".",
  mandatory = c("gene_id", "transcript_id"),
  return.obj = F,
  write.obj = T,
  do_sort = T
)
}
\arguments{
\item{gr}{Grange object}
}
\description{
Export Granges object to gtf file
}
