% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtf2bed.R
\name{gtf2bed}
\alias{gtf2bed}
\title{Convert gtf file to bed12 file.}
\usage{
gtf2bed(gtf_file, bed_file = NULL, itemRgb = "255,0,0")
}
\arguments{
\item{gtf_file}{A gtf file.}

\item{bed_file}{A bed file name to save.}
}
\value{
A bed file will be saved to the directory.
}
\description{
Convert gtf file to bed12 file.
}
