% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_gtf.R
\name{makeGtfAttributes}
\alias{makeGtfAttributes}
\title{Convert metadata information of a GRanges object to the attribute of a ftf}
\usage{
makeGtfAttributes(df, cols = NULL, mandatory = c("gene_id", "transcript_id"))
}
\arguments{
\item{df}{the metadata information of a GRanges object}
}
\description{
Convert metadata information of a GRanges object to the attribute of a ftf
}
