% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtdFilter.R
\name{rtdFilter}
\alias{rtdFilter}
\title{Filter collasped transcript datasets}
\usage{
rtdFilter(data_dir)
}
\arguments{
\item{data_dir}{Data directory}
}
\value{
Results are saved in \code{data_di}
}
\description{
Filter collasped transcript datasets
}
