% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SJanalysis.R
\name{SJanalysis}
\alias{SJanalysis}
\title{Splice Junction (SJ) analysis}
\usage{
SJanalysis(data_dir, genome_fasta = NULL, sj_overhang = 10, ref_rtd = NULL)
}
\arguments{
\item{data_dir}{The data direcotory of TAMA output}

\item{genome_fasta}{The name of genome fasta sequence file. If the file is not in the working directory, please
provide the full path.}

\item{sj_overhang}{An integer of SJ overhang. A SJ is filtered if it has mismatches at its left (-sj_overhang)
or right (+sj_overhang) exonic regions.}

\item{ref_rtd}{An gtf file of reference transcript annotation. If a canonical SJ in the isoseq assembly has 
a match in the reference transcript annotation, it will be kept even though it has mismatches in the overhangs.

genome_fasta <- 'data/isoseq/Morex_V2_pseudomolecules_and_unplaced_scaffolds_ENA.fasta'}
}
\description{
Splice Junction (SJ) analysis
}
