% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryTranSite.R
\name{SummaryTranSite}
\alias{SummaryTranSite}
\title{Summary transcript start site (TSS) and transcript end site (TES)}
\usage{
SummaryTranSite(gr, type = c("TSS", "TES"), bin = 5)
}
\arguments{
\item{gr}{A GRanges object.}

\item{type}{"TSS" for transcript start site and "TES" for transcript end site.}

\item{bin}{An integer of window size to aggregate the reads upstream and downstream of a TSS/TES.
The total reads in [site-bin,site+bin] will be calculated.}
}
\value{
A data frame of TSS or TES count.
}
\description{
Summary transcript start site (TSS) and transcript end site (TES)
}
\details{
Count the number of TSS and TES on each chromosome.
}
