% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSanalysis.R
\name{TSanalysis}
\alias{TSanalysis}
\title{Transcript start site (TSS) and end site (TES) analysis}
\usage{
TSanalysis(
  data_dir,
  cut_off = 0.05,
  cut_at = "fdr",
  TSS_region = 50,
  TES_region = 50,
  min_read = 2,
  bin = 5
)
}
\arguments{
\item{data_dir}{The data directory of TAMA output.}

\item{cut_off}{A numeric value of BH adjsuted p-value (fdr), p-value or probablity cut-off.}

\item{cut_at}{The \code{cut_off} is applied to "fdr" (adjustd p-value), "pval" (p-value) or "prob" (probability).}

\item{TSS_region, TES_region}{An integer of enriched TSS/TES wobbling region. TSSs/TESs locate within enriched
site+/-site_region are all treated as high confident sites. Default: 50.}

\item{min_read}{The minimum reads in the window around TSS/TES to support high confident sites. Default: 2.}

\item{bin}{An integer of window size to aggregate the reads upstream and downstream a TSS/TES.
The total reads in [site-bin,site+bin] will be calculated. Default: 5.}
}
\description{
Transcript start site (TSS) and end site (TES) analysis
}
