% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtdMerge.R
\name{rtdMerge}
\alias{rtdMerge}
\title{Merge inference RTD to reference RTD}
\usage{
rtdMerge(
  inf_file,
  ref_file,
  prefix_ref = "ref",
  prefix_inf = "inf",
  chimeric_tolerance = 0.05,
  data_dir
)
}
\value{
A merged RTD saved in the \code{data_di}
}
\description{
Merge inference RTD to reference RTD
}
