% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SJanalysis.R
\name{SJanalysis}
\alias{SJanalysis}
\title{Splice Junction (SJ) analysis}
\usage{
SJanalysis(data_dir, genome_fasta = NULL, sj_overhang = 10, sjdatabase = NULL)
}
\arguments{
\item{data_dir}{The data direcotory of TAMA output}

\item{genome_fasta}{The name of genome fasta sequence file. If the file is not in the working directory, please
provide the full path.}

\item{sj_overhang}{An integer of SJ overhang. A SJ is filtered if it has mismatches at its left (-sj_overhang)
or right (+sj_overhang) exonic regions.}

\item{sjdatabase}{A tab separated file of splice junction information. If a canonical SJ in the isoseq assembly has 
a match in the database, it will be kept even though it has mismatches in the overhangs.}
}
\value{
Results are saved in \code{data_di}
}
\description{
Splice Junction (SJ) analysis
}
